Update to Demand flow for both Admin and Coach paths
- Migrate demands to use proper IDs instead of array index as identifier

- Change 'cancel' to 'close'

- New demands are not open by default

- Clean up display of the demands and add coach demand view

- Add logging for debug
